# `<functional>`: make `move_only_function` do less work