--- a/benchmarks/src/move_only_function.cpp
+++ b/benchmarks/src/move_only_function.cpp
@@ -23,8 +23,8 @@ void mof_construct(benchmark::State& state) {
 void mof_move(benchmark::State& state) {
     move_only_function<void()> mof;
     for (auto _ : state) {
-        auto mof_move = move(mof);
-        benchmark::DoNotOptimize(mof_move);
+        auto moved_mof = move(mof);
+        benchmark::DoNotOptimize(moved_mof);
     }
 }
 
@@ -32,8 +32,8 @@ void mof_construct_and_move(benchmark::State& state) {
     for (auto _ : state) {
         move_only_function<void()> mof{};
         benchmark::DoNotOptimize(mof);
-        auto mof_move = move(mof);
-        benchmark::DoNotOptimize(mof_move);
+        auto moved_mof = move(mof);
+        benchmark::DoNotOptimize(moved_mof);
     }
 }