--- a/benchmarks/src/move_only_function.cpp
+++ b/benchmarks/src/move_only_function.cpp
@@ -3,6 +3,7 @@
 
 #include <benchmark/benchmark.h>
 #include <functional>
+#include <utility>
 
 using namespace std;