Primarily I treat this more as a clarity improvement.
Yet, the benchmark results prove this PR deserving `performance`  label.

Resolves #5316

# ⏱️ Benchmark results

Benchmark                |   before |    after   
-------------------------|----------|------------
mof_none                 | 0.475 ns | 0.479 ns   
mof_construct            | 0.799 ns | 0.714 ns   
mof_move                 |  1.06 ns |  1.07 ns   
mof_construct_and_move   |  1.68 ns |  1.66 ns   

# 🥇  Results interpretation

`mof_construct` is the only that changed more than results variation, it shows some improvement due to doing less work.

`mof_construct_and_move` addresses the concern of potential slowdown due to failed store-to-load forwarding. 

The concern was due to storing zero as one pointer but then loading two pointers, which can be loaded in the same vector register. If the load happens shortly after the store, it has to wait for the store to completely reach the cache, as the wider value cannot be obtained from the narrower store buffer entry.

The situation apparently does not happen, as before the actual two-pointers move, there's vtable obtaining and inspection, which takes enough time for the store to complete.

`mof_none` and `mof_move` give some idea on the corresponding baseline timings. They are not expected to change.