@StephanTLavavej 2025-03-08
Thanks! :heart_eyes_cat:

On my 5950X I actually observe `mof_construct_and_move` becoming slower:

Benchmark              | Before  | After   | Speedup
-----------------------|---------|---------|--------
mof_none               | 1.46 ns | 1.26 ns | 1.16
mof_construct          | 1.91 ns | 1.48 ns | 1.29
mof_move               | 4.21 ns | 4.18 ns | 1.01
mof_construct_and_move | 3.38 ns | 7.25 ns | 0.47

However, this case seems unrealistic (how often are users default-constructing and immediately moving), and if the code had originally been written this way, I would never have asked for an additional pointer to be nulled.