@StephanTLavavej 2025-03-08
> `mof_none` and `mof_construct_and_move` give some idea on the corresponding baseline timings. They are not expected to change.

Did you intend to say "`mof_none` and `mof_move`" here?